Put class slots back into the build ... class slot work on 2.3.8 of the GwydionDylan compiler.
Added a printing module ... it defines methods on print-object that outputs
the XML elements as XML.  I found I needed this more than once, so I added this
feature to the library.  The printing module is actually a derivative work
of the transform module.

Modified the xml-test to use the printing library of the xml-parser library.
This simplifies xml-test a bit.